Refactor Sensor builder (#1259)
